EIP1-4790 - openAPI and SQS spec changes so that IdDocumentResubmission messages include array of RejectedDocument
