+++ src/test/kotlin/uk/gov/dluhc/notificationsapi/testsupport/testdata/messaging/models/ContactDetailsBuilder.kt
@@ -1,0 +1,37 @@
+package uk.gov.dluhc.notificationsapi.testsupport.testdata.messaging.models
+
+import uk.gov.dluhc.notificationsapi.messaging.models.Address
+import uk.gov.dluhc.notificationsapi.messaging.models.ContactDetails
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.DataFaker.Companion.faker
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.getAValidPostcode
+
+fun buildContactDetailsMessage(
+    localAuthorityName: String = "some-city-council",
+    website: String = faker.company().url(),
+    phone: String = faker.phoneNumber().cellPhone(),
+    email: String = faker.internet().emailAddress(),
+    address: Address = buildAddressMessage(),
+): ContactDetails =
+    ContactDetails(
+        localAuthorityName = localAuthorityName,
+        website = website,
+        phone = phone,
+        email = email,
+        address = address
+    )
+
+fun buildAddressMessage(
+    street: String = faker.address().streetName(),
+    property: String? = faker.address().buildingNumber(),
+    locality: String? = faker.address().streetName(),
+    town: String? = faker.address().city(),
+    area: String? = faker.address().state(),
+    postcode: String = getAValidPostcode(),
+): Address = Address(
+    street = street,
+    property = property,
+    locality = locality,
+    town = town,
+    area = area,
+    postcode = postcode,
+)
