+++ src/test/kotlin/uk/gov/dluhc/notificationsapi/testsupport/testdata/messaging/models/SendNotifyPhotoResubmissionMessageBuilder.kt
@@ -1,0 +1,59 @@
+package uk.gov.dluhc.notificationsapi.testsupport.testdata.messaging.models
+
+import uk.gov.dluhc.notificationsapi.messaging.models.ContactDetails
+import uk.gov.dluhc.notificationsapi.messaging.models.Language
+import uk.gov.dluhc.notificationsapi.messaging.models.MessageAddress
+import uk.gov.dluhc.notificationsapi.messaging.models.MessageType
+import uk.gov.dluhc.notificationsapi.messaging.models.NotificationChannel
+import uk.gov.dluhc.notificationsapi.messaging.models.PhotoPersonalisation
+import uk.gov.dluhc.notificationsapi.messaging.models.PhotoRejectionReason
+import uk.gov.dluhc.notificationsapi.messaging.models.SendNotifyPhotoResubmissionMessage
+import uk.gov.dluhc.notificationsapi.messaging.models.SourceType
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.DataFaker.Companion.faker
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.aGssCode
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.aRequestor
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.aSourceReference
+import uk.gov.dluhc.notificationsapi.testsupport.testdata.aValidApplicationReference
+
+fun buildSendNotifyPhotoResubmissionMessage(
+    channel: NotificationChannel = NotificationChannel.EMAIL,
+    language: Language = Language.EN,
+    sourceType: SourceType = SourceType.VOTER_MINUS_CARD,
+    sourceReference: String = aSourceReference(),
+    gssCode: String = aGssCode(),
+    requestor: String = aRequestor(),
+    personalisation: PhotoPersonalisation = buildPhotoPersonalisationMessage(),
+    toAddress: MessageAddress = aMessageAddress(),
+): SendNotifyPhotoResubmissionMessage =
+    SendNotifyPhotoResubmissionMessage(
+        channel = channel,
+        language = language,
+        sourceType = sourceType,
+        sourceReference = sourceReference,
+        gssCode = gssCode,
+        requestor = requestor,
+        messageType = MessageType.PHOTO_MINUS_RESUBMISSION,
+        personalisation = personalisation,
+        toAddress = toAddress,
+    )
+
+fun buildPhotoPersonalisationMessage(
+    applicationReference: String = aValidApplicationReference(),
+    firstName: String = faker.name().firstName(),
+    photoRejectionReasons: List<PhotoRejectionReason> = listOf(PhotoRejectionReason.OTHER),
+    photoRejectionNotes: String? = faker.harryPotter().spell(),
+    photoRequestFreeText: String = faker.harryPotter().spell(),
+    uploadPhotoLink: String = "http://localhost:8080/eros/photo/398c1be2-7950-48a2-aca8-14cb9276a673",
+    eroContactDetails: ContactDetails = buildContactDetailsMessage()
+): PhotoPersonalisation =
+    PhotoPersonalisation(
+        applicationReference = applicationReference,
+        firstName = firstName,
+        photoRejectionReasons = photoRejectionReasons,
+        photoRejectionNotes = photoRejectionNotes,
+        photoRequestFreeText = photoRequestFreeText,
+        uploadPhotoLink = uploadPhotoLink,
+        eroContactDetails = eroContactDetails
+    )
+
+fun aSendNotifyPhotoResubmissionMessage() = buildSendNotifyPhotoResubmissionMessage()
